Extract consequence in a separated class
